fix: address review comments for and()/or() as named functions
